requirements.txt
Django==6.0
djangorestframework==3.15.2
django-jazzmin==3.0.1
asgiref==3.11.0
sqlparse==0.5.4
